Renamed NetworkLayers -> networkLayers
