+++ src/util/networkUtils.ts
